requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
colorama==0.4.6
coverage==7.6.1
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.2.2
fastapi==0.120.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.0.0
multidict==6.7.0
packaging==24.2
pluggy==1.5.0
postgrest==2.25.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==5.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.25.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
tomli==2.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
